Remove Damageable class (merge behaviours in Selectable)
